wire up rudimentary config, move pd to its own package
